Move compare-features to scripts/lib
